Remove Enter steps and always use input. Add steps at beginning to prevent step num changing
