+++ src/commands/createStaticWebApp/BuildPresetListStep.ts
@@ -3,16 +3,13 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { AzureWizardPromptStep, IAzureQuickPickItem, IWizardOptions } from 'vscode-azureextensionui';
+import { AzureWizardPromptStep, IAzureQuickPickItem } from 'vscode-azureextensionui';
 import { buildPresets } from '../../buildPresets/buildPresets';
 import { IBuildPreset } from '../../buildPresets/IBuildPreset';
 import { ext } from '../../extensionVariables';
 import { localize } from '../../utils/localize';
 import { openUrl } from '../../utils/openUrl';
-import { ApiLocationStep } from './ApiLocationStep';
-import { AppLocationStep } from './AppLocationStep';
 import { IStaticWebAppWizardContext } from './IStaticWebAppWizardContext';
-import { OutputLocationStep } from './OutputLocationStep';
 
 export class BuildPresetListStep extends AzureWizardPromptStep<IStaticWebAppWizardContext> {
 
@@ -44,13 +41,5 @@ export class BuildPresetListStep extends AzureWizardPromptStep<IStaticWebAppWiza
 
     public shouldPrompt(context: IStaticWebAppWizardContext): boolean {
         return !context.appLocation;
-    }
-
-    public async getSubWizard(context: IStaticWebAppWizardContext): Promise<IWizardOptions<IStaticWebAppWizardContext> | undefined> {
-        if (!context.appLocation) {
-            return { promptSteps: [new AppLocationStep(), new ApiLocationStep(), new OutputLocationStep()] };
-        } else {
-            return undefined;
-        }
     }
 }
